fix: typechecker working with promotions and not allowing int type mismatch

+++ input/test_i64.rs
@@ -1,6 +1,6 @@
 fn main() {
     let mut a: i64 = 10;
-    let b: i32 = 5;
+    let b: i64 = 5;
     let c = a + b; // mixing i64 and i32
     println!("{}", c);
     let d = b + 3; // constant + variable
